default process added and screen text translated to english

+++ src/app/ktadmin/alternativa/alternativa.component.ts
@@ -4,11 +4,12 @@ import { LocalDataSource } from 'ng2-smart-table';
 import { AlternativaService } from './alternativa.service';
 import { AlternativaKtadmin } from '../model/alternativa_ktadmin';
 import { BotaoPossuirProcessosComponent } from './botao-possuir-processos/botao-possuir-processos.component';
-import { NbWindowService } from '@nebular/theme';
+import { NbGlobalPhysicalPosition, NbGlobalPosition, NbToastrService, NbWindowService } from '@nebular/theme';
 import { ProcessosPossuidosComponent } from './processos-possuidos/processos-possuidos.component';
 
 import { Subscription } from 'rxjs';
 import { AlternativaPost } from '../model/alternativa_post';
+
 
 @Component({
   selector: 'ngx-alternativa',
@@ -39,17 +40,17 @@ export class AlternativaComponent{
     },    
     columns: {
       alternativa_id: {
-        title: 'Alternativa ID',
+        title: 'Alternative ID',
         type: 'number',
         editable: false,
         addable:false
       },
       pergunta: {
-        title: 'Pergunta',
+        title: 'Alternative',
         type: 'string',
       },
        processos: {
-         title: 'Processos',
+         title: 'Process',
          type:'custom',
          editable: false,
          addable:false,
@@ -69,12 +70,15 @@ export class AlternativaComponent{
   source: LocalDataSource = new LocalDataSource();
   constructor(
     private alternativa_service: AlternativaService,
-    private windowService: NbWindowService
+    private windowService: NbWindowService,
+    private toastrService: NbToastrService
     ) {
       this.alternativas=[];
-      this.source.load(this.alternativas);
+      //this.source.load(this.alternativas);
+      
        this.alternativa_service.getAlternativas().subscribe(response =>{
-         this.alternativas=response;
+        this.alternativas=response
+         //this.alternativas.unshift(globalDataMock.defaultAlternativa);
          this.source.load(this.alternativas);
        });
   }
@@ -97,14 +101,45 @@ export class AlternativaComponent{
   }
 
   onEditConfirm(event): void {
-    this.alternativa_service.updateAlternativa(event.newData).subscribe(event.confirm.resolve());
+    if(event.data.alternativa_id==-1)this.showToast("Default Alternative can't be eddited");
+    else
+    {
+      this.alternativa_service.updateAlternativa(event.newData).subscribe(event.confirm.resolve());
+    }
   }
 
+
   onDeleteConfirm(event): void {
-    if (window.confirm('Are you sure you want to delete?')) {
-      this.alternativa_service.deleteAlternativa(event.data.alternativa_id).subscribe(event.confirm.resolve());
-    } else {
-      event.confirm.reject();
+    if(event.data.alternativa_id==-1)
+      {
+        this.showToast("Default Alternative can't be deleted");
+        event.confirm.resolve();
+      }
+    else
+    {
+      if (window.confirm('Are you sure you want to delete?')) {
+        this.alternativa_service.deleteAlternativa(event.data.alternativa_id).subscribe(event.confirm.resolve());
+      } else {
+        event.confirm.reject();
+      }
     }
+  }
+  position: NbGlobalPosition = NbGlobalPhysicalPosition.TOP_RIGHT;
+
+  private showToast(message:string) {
+    const config = {
+      status: 'warning',
+      destroyByClick: true,
+      duration: 2000,
+      hasIcon: true,
+      position: this.position,
+      preventDuplicates: false
+    };
+
+    this.toastrService.show(
+      message,
+      //"Default Alternative can't be deleted",
+      `Warning`,
+      config);
   }
 }
